+++ api/src/main/java/com/api/bank/model/entity/Operation.java
@@ -22,7 +22,7 @@ public class Operation extends Base {
     private String label;
 
     @Column(nullable = false)
-    private float amount;
+    private double amount;
 
     private Date operationDate;
 
@@ -49,7 +49,7 @@ public class Operation extends Base {
     }
 
     public Operation(String operationId,
-                     String label, float amount,
+                     String label, double amount,
                      Date operationDate,
                      Account account, QrCheck qrCheck, OperationStatus operationStatus,
                      OperationType operationType, PaymentMethod paymentMethod) {
